requirements.txt
pandas
numpy
OpenNMT-py
youtube_transcript_api
python-dotenv
google-api-python-client
google-cloud-speech
scipy
nltk
indic-nlp-library
SpeechRecognition
pydub
ffmpeg
youtube_dl
ffprobe
inltk
sentencepiece
oauth2client
sacrebleu